+++ src/main/kotlin/kr/hhplus/be/server/domain/concert/aop/IncrementViewCount.kt
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.domain.concert.aop
+
+@Target(AnnotationTarget.FUNCTION)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class IncrementViewCount(
+    val concertIdParam: String = "concertId"
+)
